Changed default Ordering and ignore lambdas
Now we ignore lambda functions because usually they're not useful when
reversing a game.

Also, changed all the ordering to Ordering::Relaxed since we were using
it mostly for statistics which are not critical sections.

Also, changed all offsets to relative offsetting because the addresses
are not persistant on game launch because of ASRL.